Added exceptions and set a loading screen on select seat

+++ lib/screens/SelectSeatRoundWay.dart
@@ -18,6 +18,7 @@ class SelectSeat_RoundWay extends StatefulWidget {
 }
 
 class _SelectSeat_RoundWayState extends State<SelectSeat_RoundWay> {
+  Function loadScreen;
   SharePreferencelogin sh = SharePreferencelogin();
   String selectTicket = "Select a Seat";
   bool unblockLoad;
@@ -43,7 +44,13 @@ class _SelectSeat_RoundWayState extends State<SelectSeat_RoundWay> {
   void initState() {
     super.initState();
     selectTicketlan();
-
+    loadScreen = () {
+      if (unblockLoad)
+        unblockLoad = false;
+      else
+        unblockLoad = true;
+      setState(() {});
+    };
     unblockLoad = false;
   }
 
@@ -51,7 +58,7 @@ class _SelectSeat_RoundWayState extends State<SelectSeat_RoundWay> {
   Widget build(BuildContext context) {
     return WillPopScope(
         onWillPop: () async {
-          if (selectedSeats.length > 0) {
+          if (selectedSeats.length > 0 && !unblockLoad) {
             List pointNumbers = selectedSeats.keys.toList();
             SharedPreferences prefs = await SharedPreferences.getInstance();
             String token = prefs.getString('Token');
@@ -64,8 +71,9 @@ class _SelectSeat_RoundWayState extends State<SelectSeat_RoundWay> {
               );
             }
             selectedSeats.clear();
+            unblockLoad = false;
           }
-          Navigator.pop(context);
+          if (!unblockLoad) Navigator.pop(context);
         },
         child: GunselScaffold(
           appBarIcon: backArrow,
@@ -75,6 +83,7 @@ class _SelectSeat_RoundWayState extends State<SelectSeat_RoundWay> {
             children: <Widget>[
               SelectSeatScreen(
                 ticketData: this.widget.ticketData,
+                loadScreenFunction: loadScreen,
               ),
               unblockLoad
                   ? Container(
@@ -89,7 +98,7 @@ class _SelectSeat_RoundWayState extends State<SelectSeat_RoundWay> {
             ],
           ),
           customFunction: () async {
-            if (selectedSeats.length > 0) {
+            if (selectedSeats.length > 0 && !unblockLoad) {
               List pointNumbers = selectedSeats.keys.toList();
               SharedPreferences prefs = await SharedPreferences.getInstance();
               String token = prefs.getString('Token');
@@ -100,11 +109,11 @@ class _SelectSeat_RoundWayState extends State<SelectSeat_RoundWay> {
                   'https://test-api.gunsel.ua/Public.svc/UnblockTravelSeat/${selectedSeats[pointNumber]}',
                   headers: {'token': token},
                 );
-                print(response.body);
               }
               selectedSeats.clear();
+              unblockLoad = false;
             }
-            Navigator.of(context).pop();
+            if (!unblockLoad) Navigator.of(context).pop();
           },
           appBarTitle: selectTicket,
           appBarTitleIncluded: true,
@@ -115,8 +124,10 @@ class _SelectSeat_RoundWayState extends State<SelectSeat_RoundWay> {
 
 class SelectSeatScreen extends StatefulWidget {
   Map<String, dynamic> ticketData;
+  Function loadScreenFunction;
   SelectSeatScreen({
     @required this.ticketData,
+    @required this.loadScreenFunction,
   });
 
   @override
@@ -132,7 +143,6 @@ class _SelectSeatScreenState extends State<SelectSeatScreen> {
   @override
   void initState() {
     selectTicketlan();
-    debugPrint(widget.ticketData['FirstLeg'].toString(), wrapWidth: 1024);
     getData();
     super.initState();
     _dataFetched = getData();
@@ -442,6 +452,7 @@ class _SelectSeatScreenState extends State<SelectSeatScreen> {
                           ? getSpacer()
                           : Container(),
                       Seat(
+                        loadScreen: widget.loadScreenFunction,
                         refresh: this.refresh,
                         seatData: seatMap[(index + 1)],
                         ticketData: widget.ticketData,
@@ -464,11 +475,13 @@ class _SelectSeatScreenState extends State<SelectSeatScreen> {
 class Seat extends StatefulWidget {
   Map<String, dynamic> seatData;
   Function refresh;
+  Function loadScreen;
   Map<String, dynamic> ticketData;
   Seat({
     @required this.refresh,
     @required this.seatData,
     @required this.ticketData,
+    @required this.loadScreen,
   });
   @override
   _SeatState createState() => _SeatState();
@@ -509,6 +522,7 @@ class _SeatState extends State<Seat> {
               });
               String url =
                   'https://test-api.gunsel.ua/Public.svc/BlockTravelSeat';
+              widget.loadScreen();
               http.Response response = await http.post(
                 url,
                 body: body,
@@ -525,11 +539,13 @@ class _SeatState extends State<Seat> {
                     widget.seatData['TravelSeatBlockId'];
 
                 widget.refresh();
+                widget.loadScreen();
                 setState(
                   () {},
                 );
               }
-            } else {
+            } else if (makeItGreen) {
+              makeItGreen = false;
               http.Response response = await http.delete(
                 'https://test-api.gunsel.ua/Public.svc/UnblockTravelSeat/${widget.seatData['TravelSeatBlockId']}',
                 headers: {'token': token},
@@ -537,9 +553,8 @@ class _SeatState extends State<Seat> {
               widget.seatData['TravelSeatBlockId'] = null;
               selectedSeats.remove(widget.seatData['PointNumber']);
               widget.refresh();
-              setState(() {
-                makeItGreen = false;
-              });
+              widget.loadScreen();
+              setState(() {});
             }
           }
         }
